+++ common/src/main/java/io/github/mortuusars/scholar/client/gui/screen/edit/LecternSpreadBookEditScreen.java
@@ -13,7 +13,6 @@ import net.minecraft.client.gui.components.events.GuiEventListener;
 import net.minecraft.client.gui.screens.inventory.MenuAccess;
 import net.minecraft.network.chat.CommonComponents;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.*;
 import net.minecraft.world.item.ItemStack;
@@ -40,7 +39,7 @@ public class LecternSpreadBookEditScreen extends SpreadBookEditScreen implements
     };
 
     public LecternSpreadBookEditScreen(LecternSpreadBookEditMenu lecternSpreadMenu, Inventory inventory, Component component) {
-        super(lecternSpreadMenu.getBook(), InteractionHand.MAIN_HAND);
+        super(lecternSpreadMenu.getBook());
         this.menu = lecternSpreadMenu;
     }
 
@@ -76,7 +75,9 @@ public class LecternSpreadBookEditScreen extends SpreadBookEditScreen implements
     protected boolean pageBack() {
         this.sendButtonClick(LecternMenu.BUTTON_PREV_PAGE);
 
-        getHistory().add(() -> this.sendButtonClick(LecternMenu.BUTTON_PREV_PAGE), () -> this.sendButtonClick(LecternMenu.BUTTON_NEXT_PAGE)
+        getHistory().add(
+              () -> this.sendButtonClick(LecternMenu.BUTTON_PREV_PAGE),
+              () -> this.sendButtonClick(LecternMenu.BUTTON_NEXT_PAGE)
         );
 
         return true;
@@ -86,7 +87,9 @@ public class LecternSpreadBookEditScreen extends SpreadBookEditScreen implements
     protected boolean pageForward() {
         this.sendButtonClick(LecternMenu.BUTTON_NEXT_PAGE);
 
-        getHistory().add(() -> this.sendButtonClick(LecternMenu.BUTTON_NEXT_PAGE), () -> this.sendButtonClick(LecternMenu.BUTTON_PREV_PAGE)
+        getHistory().add(
+              () -> this.sendButtonClick(LecternMenu.BUTTON_NEXT_PAGE),
+              () -> this.sendButtonClick(LecternMenu.BUTTON_PREV_PAGE)
         );
 
         return true;
@@ -108,7 +111,7 @@ public class LecternSpreadBookEditScreen extends SpreadBookEditScreen implements
             }
 
             setTextBoxes();
-            updateButtonVisibility();
+            updateButtons();
         }
         saveChanges(false, null);
     }
@@ -118,11 +121,6 @@ public class LecternSpreadBookEditScreen extends SpreadBookEditScreen implements
         removeEmptyTrailingPages();
         // Copying 'pages' list to not cause ConcurrentModificationException when packet is encoded:
         Packets.sendToServer(new LecternEditBookC2SP(getMenu().getLecternPos(), new ArrayList<>(pages), Optional.ofNullable(title)));
-    }
-
-    @Override
-    protected void updateLocalCopy() {
-        super.updateLocalCopy();
     }
 
     // --
@@ -171,7 +169,6 @@ public class LecternSpreadBookEditScreen extends SpreadBookEditScreen implements
     }
 
     // --
-
 
     @Override
     public void setFocused(@Nullable GuiEventListener focused) {