+++ common/src/main/java/io/github/mortuusars/scholar/mixin/HandleOpenBookClientPacketListenerMixin.java
@@ -3,7 +3,7 @@ package io.github.mortuusars.scholar.mixin;
 import io.github.mortuusars.scholar.Config;
 import io.github.mortuusars.scholar.client.gui.screen.view.BookViewAccess;
 import io.github.mortuusars.scholar.book.BookColor;
-import io.github.mortuusars.scholar.client.gui.screen.view.SpreadBookViewScreen;
+import io.github.mortuusars.scholar.client.gui.screen.view.InHandSpreadBookViewScreen;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientPacketListener;
 import net.minecraft.network.protocol.game.ClientboundOpenBookPacket;
@@ -20,17 +20,21 @@ public abstract class HandleOpenBookClientPacketListenerMixin {
     @Inject(method = "handleOpenBook", at = @At(value = "INVOKE",
             target = "Lnet/minecraft/client/player/LocalPlayer;getItemInHand(Lnet/minecraft/world/InteractionHand;)Lnet/minecraft/world/item/ItemStack;"),
             cancellable = true)
-    private void handleOpenBook(ClientboundOpenBookPacket clientboundOpenBookPacket, CallbackInfo ci) {
+    private void handleOpenBook(ClientboundOpenBookPacket packet, CallbackInfo ci) {
         if (Minecraft.getInstance().player == null) return;
         if (!Config.Common.IN_HAND_TWO_PAGE_BOOK_SCREEN.get()) return;
         if (Config.Common.SNEAK_OPENS_VANILLA_BOOK_SCREEN.get() && Minecraft.getInstance().player.isSecondaryUseActive()) return;
 
-        InteractionHand hand = clientboundOpenBookPacket.getHand();
+        InteractionHand hand = packet.getHand();
         ItemStack stack = Minecraft.getInstance().player.getItemInHand(hand);
 
         if (!stack.is(Items.WRITTEN_BOOK)) return;
 
-        Minecraft.getInstance().setScreen(new SpreadBookViewScreen(BookViewAccess.fromItem(stack), BookColor.of(stack)));
+        Minecraft.getInstance().setScreen(new InHandSpreadBookViewScreen(
+              BookViewAccess.fromItem(stack),
+              BookColor.of(stack),
+              packet.getHand()));
+
         ci.cancel();
     }
 }