+++ common/src/main/java/io/github/mortuusars/scholar/network/packet/C2SPackets.java
@@ -1,6 +1,7 @@
 package io.github.mortuusars.scholar.network.packet;
 
 import io.github.mortuusars.scholar.network.packet.server.LecternEditBookC2SP;
+import io.github.mortuusars.scholar.network.packet.server.SetBookmarkC2SP;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 
@@ -9,7 +10,8 @@ import java.util.List;
 public class C2SPackets {
     public static List<CustomPacketPayload.TypeAndCodec<? extends FriendlyByteBuf, ? extends CustomPacketPayload>> getDefinitions() {
         return List.of(
-                new CustomPacketPayload.TypeAndCodec<>(LecternEditBookC2SP.TYPE, LecternEditBookC2SP.STREAM_CODEC)
+                new CustomPacketPayload.TypeAndCodec<>(LecternEditBookC2SP.TYPE, LecternEditBookC2SP.STREAM_CODEC),
+                new CustomPacketPayload.TypeAndCodec<>(SetBookmarkC2SP.TYPE, SetBookmarkC2SP.STREAM_CODEC)
         );
     }
 }